Add watch option to Verify command
